+++ Tests/ColumbusTests.swift
@@ -3,7 +3,7 @@
 //  Columbus
 //
 //  Created by Stefan Herold on 24.12.18.
-//  Copyright © 2022 Stefan Herold. All rights reserved.
+//  Copyright © 2023 Stefan Herold. All rights reserved.
 //
 
 import XCTest