+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 * Implement DisplayState - [@Blackjacx](https://github.com/Blackjacx).
 * Inject config directly in ViewController - [@Blackjacx](https://github.com/Blackjacx).
 
+## [1.9.0] - 2023-05-09Z
+* [#49](https://github.com/Blackjacx/Columbus/pull/49): Add fallback for CTCarrier deprecation - [@mic-ioki](https://github.com/mic-ioki).
+* Use default country for the initial country code - [@Blackjacx](https://github.com/Blackjacx).
+* Update Copyright - [@Blackjacx](https://github.com/Blackjacx).
+
 ## [1.8.0] - 2023-05-02Z
 * Remove auto assign to project workflow - [@Blackjacx](https://github.com/Blackjacx).
 * Make CoreTelephony Approach Debuggable - [@Blackjacx](https://github.com/Blackjacx).